Sync desktop-icons patches for master
The next rc seems a little way off although the repo has definitely
stabilised. Let's sync the patches to fix some annoying common bugs
seen in the extension, such as handling of .desktop file rendering
and labels.

This allows us to drop the custom patch we had before for system-wide
installation and the configuration option as non-standard, rebasing
on vanilla upstream.

Deliberately omitting binary git diff to remove compiled schemas
as it will not apply in mock.

Signed-off-by: Ikey Doherty <ikey.doherty@intel.com>